+++ rd/trunk/Client/cms/Assets/script/SpellService/EffectSet.cs
@@ -25,7 +25,7 @@ public class EffectSet : Effect
         base.Init(pt, owner);
     }
     //---------------------------------------------------------------------------------------------
-    public override void Apply(int applyTime, float aniDelayTime)
+    public override void Apply(float applyTime, float aniDelayTime)
     {
         base.Apply(applyTime);
